MAINT: Rename `classic_models` by `derived_models`
