+++ src/porepy/applications/verification_setups/terzaghi_biot.py
@@ -45,7 +45,7 @@ import numpy as np
 import porepy as pp
 import porepy.models.fluid_mass_balance as mass
 import porepy.models.poromechanics as poromechanics
-from porepy.applications.classic_models.biot import BiotPoromechanics
+from porepy.applications.derived_models.biot import BiotPoromechanics
 from porepy.applications.verification_setups.verification_utils import VerificationUtils
 
 number = pp.number
